Fix codestyle
The current script failed to install the code style. Because of the newer folders with spaces in their name (Application\ Support), the script was creating a new directory called "Support" in the root of the project.

I guess this must be happening since some MacOS version or something. But I fixed it based on this commit from the original repository: https://github.com/square/java-code-styles/commit/edfc51e801bf538d661be5484be684bbc3a786df