Avoid context parameter to enable Overlay
The `enableOverlay(context)` behaves erratically when passing different types of context.
If the Context is an Activity, it enables the Overlay in the current screen.
If it's other type (application, ContextThemeWrapper) it won't enable it on the current screen, but will enable it on the next screen you open. This is weird.

To make the API easier to use, I removed the `context` parameter from the `enableOverlay` function. Instead, I kept a WeakReference of the current Activity. Whenever the overlay is enabled, we inject the overlay on this activity (if exists). That way we can safely enable the overlay from anywhere.